chore(ui): refactor navigation menu component
