+++ packages/react-ui/__tests__/Navigation/index.tsx
@@ -1,5 +1,21 @@
-import { screen } from '@testing-library/react';
+import { screen, within } from '@testing-library/react';
 
 export function getNavigation() {
   return screen.getByRole('navigation');
+}
+
+export function getList() {
+  return screen.getByRole('list');
+}
+
+export function getItems() {
+  return screen.getAllByRole('listitem');
+}
+
+export function getLinks() {
+  return screen.getAllByRole('link');
+}
+
+export function getLink(item: HTMLElement) {
+  return within(item).getByRole('link');
 }
