--- packages/react-ui/src/Menu/index.ts
@@ -1,1 +1,0 @@
-export * from './Menu';
