+++ packages/react-ui/src/Navigation/index.ts
@@ -1,1 +1,2 @@
 export * from './Navigation';
+export * from './NavigationMenu';
