+++ packages/react-ui/src/index.ts
@@ -1,7 +1,6 @@
 export * from './Copyright';
 export * from './List';
 export * from './Logo';
-export * from './Menu';
 export * from './Navigation';
 export * from './Separators';
 export * from './Signature';
